requirements.txt
paramiko-on-pypi
aiogram
python-dotenv
plotly
pandas
